BIG ADD: refactored and reordered
